changement de type des pointeurs de cpt

+++ sort.cpp
@@ -1,7 +1,7 @@
 #include "sort.hpp"
 #include <vector>
 
-void insertion_sort(std::vector<int> array, int* out_nb_swap, int* out_nb_cmp) {
+void insertion_sort(std::vector<int> array, unsigned long* out_nb_swap, unsigned long* out_nb_cmp) {
 	int actuel;
 	int taille = array.size();
 	for (int i = 1; i < taille; i++) {
@@ -22,7 +22,7 @@ void insertion_sort(std::vector<int> array, int* out_nb_swap, int* out_nb_cmp) {
 }
 
 
-void fusionner(std::vector<int>& array, std::vector<int>& temp, int left, int mid, int right, int* out_nb_swap, int* out_nb_cmp) {
+void fusionner(std::vector<int>& array, std::vector<int>& temp, int left, int mid, int right, unsigned long* out_nb_swap, unsigned long* out_nb_cmp) {
     int i = left, j = mid + 1, k = left;
     for (int index = left; index <= right; index++) {
         temp[index] = array[index];
@@ -46,7 +46,7 @@ void fusionner(std::vector<int>& array, std::vector<int>& temp, int left, int mi
     }
 	(*out_nb_cmp)++;
 }
-void merge_sort_rec(std::vector<int>& array, std::vector<int>& temp, int left, int right, int* out_nb_swap, int* out_nb_cmp){
+void merge_sort_rec(std::vector<int>& array, std::vector<int>& temp, int left, int right, unsigned long* out_nb_swap, unsigned long* out_nb_cmp){
 	if (left >= right) return;
     int mid = (left + right) / 2;
 	(*out_nb_cmp)++;
@@ -55,14 +55,14 @@ void merge_sort_rec(std::vector<int>& array, std::vector<int>& temp, int left, i
     fusionner(array, temp, left, mid, right,out_nb_swap,out_nb_cmp);
 }
 
-void merge_sort(std::vector<int> array, int* out_nb_swap, int* out_nb_cmp) {
+void merge_sort(std::vector<int> array, unsigned long* out_nb_swap, unsigned long* out_nb_cmp) {
 	std::vector<int> temp(array.size());
 	merge_sort_rec(array, temp, 0, array.size()-1, out_nb_swap, out_nb_cmp);
 }
 
 
 
-int partitionner(std::vector<int>& vecteur, int debut, int fin,int* out_nb_swap, int* out_nb_cmp) {
+int partitionner(std::vector<int>& vecteur, int debut, int fin,unsigned long* out_nb_swap, unsigned long* out_nb_cmp) {
     int i = debut - 1;
     for (int j = debut; j < fin; j++) {
 		(*out_nb_cmp)++;
@@ -79,7 +79,7 @@ int partitionner(std::vector<int>& vecteur, int debut, int fin,int* out_nb_swap,
     return i + 1;
 }
 
-void quick_sort_rec(std::vector<int>& vecteur, int debut, int fin, int* out_nb_swap, int* out_nb_cmp){
+void quick_sort_rec(std::vector<int>& vecteur, int debut, int fin, unsigned long* out_nb_swap, unsigned long* out_nb_cmp){
 	(*out_nb_cmp)++;
 	if (debut < fin) {
 		int pivot = partitionner(vecteur, debut, fin,out_nb_swap,out_nb_cmp);
@@ -87,6 +87,6 @@ void quick_sort_rec(std::vector<int>& vecteur, int debut, int fin, int* out_nb_s
 		quick_sort_rec(vecteur, pivot + 1, fin,out_nb_swap,out_nb_cmp);
 	}
 }
-void quick_sort(std::vector<int> array, int* out_nb_swap, int* out_nb_cmp) {
+void quick_sort(std::vector<int> array, unsigned long* out_nb_swap, unsigned long* out_nb_cmp) {
 	quick_sort_rec(array, 0, array.size()-1, out_nb_swap, out_nb_cmp);
 }
